Create a skeleton app

This function creates a skeleton app containing empty modules

init(
  path,
  name,
  include_map,
  include_table,
  include_code,
  common_objects,
  modules,
  author
)

Arguments

path

The path to where the app should be created

name

The name of the app which will be used as the package name

include_map

Whether to include a leaflet map

include_table

Whether to include a table tab

include_code

Whether to include a tab for viewing module code

common_objects

A vector of the objects which will be shared between modules

modules

A dataframe containing long and short names of components (tabs) and modules in the order to be included and whether they should include mapping, save, markdown and result functionality. The component and module columns are used to generate file names for the modules. The long_component and long_module columns are used to generate UI and so should be formatted appropriately.

author

The name of the author(s)

Author

Simon E. H. Smart <simon.smart@cantab.net>

Examples

modules <- data.frame(
"component" = c("data", "data", "plot", "plot"),
"long_component" = c("Load data", "Load data", "Plot data", "Plot data"),
"module" = c("user", "database", "histogram", "scatter"),
"long_module" = c("Upload your own data", "Query a database to obtain data",
"Plot the data as a histogram", "Plot the data as a scatterplot"),
"map" = c(TRUE, TRUE, FALSE, FALSE),
"result" = c(FALSE, FALSE, TRUE, TRUE),
"rmd" = c(TRUE, TRUE, TRUE, TRUE),
"save" = c(TRUE, TRUE, TRUE, TRUE))
common_objects = c("raster", "histogram", "scatter")
init(path = "~/Documents", name = "demo_app",
include_map = TRUE, include_table = TRUE, include_code = TRUE,
common_objects = common_objects, modules = modules,
author = "Simon E. H. Smart")
#> Warning: cannot create dir '/home/runner/Documents/demo_app', reason 'No such file or directory'
#> Warning: cannot create dir '/home/runner/Documents/demo_app/R', reason 'No such file or directory'
#> 
#> 
#> processing file: /tmp/RtmpWrVOpK/file245c5bc25962.Rmd
#> 1/9 [unnamed-chunk-1]
#> 2/9                  
#> 3/9 [unnamed-chunk-2]
#> 4/9                  
#> 5/9 [unnamed-chunk-3]
#> 6/9                  
#> 7/9 [unnamed-chunk-4]
#> 8/9 [unnamed-chunk-5]
#> 9/9 [unnamed-chunk-6]
#> output file: ~/Documents/demo_app/inst/shiny/server.R
#> 
#> 
#> processing file: /tmp/RtmpWrVOpK/file245c5359c31e.Rmd
#> 1/7 [unnamed-chunk-1]
#> 2/7 [unnamed-chunk-2]
#> 3/7 [unnamed-chunk-3]
#> 4/7 [unnamed-chunk-4]
#> 5/7 [unnamed-chunk-5]
#> 6/7 [unnamed-chunk-6]
#> 7/7                  
#> output file: ~/Documents/demo_app/inst/shiny/ui.R
#> 
#> 
#> processing file: /tmp/RtmpWrVOpK/file245c3ece10f7.Rmd
#> 1/2 [unnamed-chunk-1]
#> 2/2                  
#> output file: ~/Documents/demo_app/inst/shiny/global.R
#> Warning: cannot open file '/home/runner/Documents/demo_app/R/data_user.R': No such file or directory
#> Error in file(con, "w"): cannot open the connection